pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gmail-mail-client"
version = "0.1.0"
description = "Simple Gmail SMTP client library with OAuth2"
dependencies = [
    "pydantic[email]>=2.0.0",
    "google-auth>=2.0.0",
    "google-auth-oauthlib>=1.0.0",
    "google-auth-httplib2>=0.1.0",
    "google-api-python-client>=2.0.0",
]

[project.scripts]
gmail-send = "mail.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
]